Move CenteredContent to separate component
