+++ src/components/Favorites/Favorites.jsx
@@ -4,8 +4,8 @@ import {
   getFromStorageData,
   setToStorageData,
 } from '../../utilities/localStorage';
+import CenteredContent from '../UI/CenteredContent/CenteredContent';
 import GridLayout from '../UI/Grid/GridLayout/GridLayout';
-import { CenteredContent } from '../UI/Grid/GridRow/styles';
 import SectionTitle from '../UI/SectionTitle/SectionTitle';
 
 import FavoritesList from './FavoritesList/FavoritesList';
