+++ src/components/UI/CenteredContent/CenteredContent.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { CenteredContentWrapper } from './styles';
+
+const CenteredContent = ({ children }) => (
+  <CenteredContentWrapper>{children}</CenteredContentWrapper>
+);
+
+export default CenteredContent;
