+++ src/components/UI/CenteredContent/styles.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const CenteredContentWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
