+++ src/components/UI/Grid/GridRow/styles.js
@@ -57,9 +57,3 @@ export const PrimaryContent = styled.div`
 export const SecondaryContent = styled.div`
   color: ${({ theme }) => theme.neutralColor};
 `;
-
-export const CenteredContent = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
